fix: id function
