+++ src/mappings/handlers/index.ts
@@ -1,4 +1,4 @@
 export { handleMultisigCall, handleMultisigInProxy } from "./multisigCallHandler";
-export { handlePureProxyEvent } from "./pureProxyEventHandler";
+export { handlePureProxyEvent, handlePureProxyKiledEvent } from "./pureProxyEventHandler";
 export { handleMultisigRemarkEventHandler } from "./multisigRemarkHandler";
-export { handleProxyEvent, handleProxyRemovedEvent, handlePureProxyKiledEvent } from "./proxyEventHandler";
+export { handleProxyEvent, handleProxyRemovedEvent } from "./proxyEventHandler";
