+++ src/utils/extractProxyEventData.ts
@@ -3,6 +3,10 @@ import { SubstrateEvent } from "@subql/types";
 import { decodeAddress } from "./addressesDecode";
 import { HexString } from "@polkadot/util/types";
 
+export function getProxiedId({ chainId, accountId, proxyAccountId, type, delay }: { chainId: string, accountId: string, proxyAccountId: string, type: string, delay: number }) {
+  return `${chainId}-${accountId}-${proxyAccountId}-${type}-${delay}`;
+}
+
 function extrinsicIndex(event: SubstrateEvent): number {
   return event.extrinsic ? event.extrinsic.idx : event.idx;
 }
